+++ R/export_table.R
@@ -47,6 +47,7 @@
 #'   output, like markdown files), the table is split into two parts. Else,
 #'   if `table_width` is numeric and table rows are larger than `table_width`,
 #'   the table is split into two parts.
+#' @param ... Currently not used.
 #' @inheritParams format_value
 #' @inheritParams get_data
 #'
@@ -118,7 +119,8 @@ export_table <- function(x,
                          group_by = NULL,
                          zap_small = FALSE,
                          table_width = NULL,
-                         verbose = TRUE) {
+                         verbose = TRUE,
+                         ...) {
   # check args
   if (is.null(format)) {
     format <- "text"
@@ -481,7 +483,7 @@ print.insight_table <- function(x, ...) {
   # been done previously, we now look for column-specific alignments
 
   if (!is.null(align) && length(align) == 1) {
-    for (i in 1:ncol(final)) {
+    for (i in seq_len(ncol(final))) {
       align_char <- ""
       if (align %in% c("left", "right", "center", "firstleft")) {
         align_char <- ""
@@ -656,7 +658,7 @@ print.insight_table <- function(x, ...) {
 .table_parts <- function(rows, final, header, sep, cross, empty_line) {
   # "final" is a matrix here. we now paste each row into a character string,
   # add separator chars etc.
-  for (row in 1:nrow(final)) {
+  for (row in seq_len(nrow(final))) {
     # create a string for each row, where cells from original matrix are
     # separated by the separator char
     final_row <- paste0(final[row, ], collapse = sep)
@@ -765,7 +767,7 @@ print.insight_table <- function(x, ...) {
   final[grp_rows, 1] <- paste0(whitespace, final[grp_rows, 1])
 
   # find rows that should not be indented
-  non_grp_rows <- 1:nrow(final)
+  non_grp_rows <- seq_len(nrow(final))
   non_grp_rows <- non_grp_rows[!non_grp_rows %in% grp_rows]
 
   # paste whitespace at end, to ensure equal width for each string
@@ -790,7 +792,7 @@ print.insight_table <- function(x, ...) {
   final[grp_rows, 1] <- paste0(whitespace, final[grp_rows, 1])
 
   # find rows that should not be indented
-  non_grp_rows <- 1:nrow(final)
+  non_grp_rows <- seq_len(nrow(final))
   non_grp_rows <- non_grp_rows[!non_grp_rows %in% grp_rows]
 
   # remove indent token
@@ -881,7 +883,7 @@ print.insight_table <- function(x, ...) {
 
   # Transform to character
   rows <- c()
-  for (row in 1:nrow(final)) {
+  for (row in seq_len(nrow(final))) {
     final_row <- paste0("|", paste0(final[row, ], collapse = "|"), "|", collapse = "")
     rows <- c(rows, final_row)