+++ R/find_formula.R
@@ -123,7 +123,7 @@ find_formula.list <- function(x, verbose = TRUE, ...) {
 
 #' @export
 find_formula.data.frame <- function(x, verbose = TRUE, ...) {
-  stop("A data frame is not a valid object for this function.")
+  stop("A data frame is not a valid object for this function.", call. = FALSE)
 }
 
 
@@ -137,7 +137,7 @@ find_formula.aovlist <- function(x, verbose = TRUE, ...) {
 
 #' @export
 find_formula.anova <- function(x, verbose = TRUE, ...) {
-  stop("Formulas cannot be retrieved from anova() objects.")
+  stop("Formulas cannot be retrieved from anova() objects.", call. = FALSE)
 }
 
 
@@ -1721,11 +1721,12 @@ find_formula.model_fit <- function(x, verbose = TRUE, ...) {
 
   if (grepl("(.*)poly\\((.*),\\s*raw\\s*=\\s*T\\)", f)) {
     if (verbose) {
-      warning(format_message(
-        "Looks like you are using 'poly()' with 'raw = T'. This results in unexpected behaviour, because 'all.vars()' considers 'T' as variable.",
-        "Please use 'raw = TRUE'."
-      ),
-      call. = FALSE
+      warning(
+        format_message(
+          "Looks like you are using 'poly()' with 'raw = T'. This results in unexpected behaviour, because 'all.vars()' considers 'T' as variable.",
+          "Please use 'raw = TRUE'."
+        ),
+        call. = FALSE
       )
     }
     return(FALSE)