+++ R/get_residuals.R
@@ -116,10 +116,12 @@ get_residuals.default <- function(x, weighted = FALSE, verbose = TRUE, ...) {
   }
 
   if (is.null(res) || all(is.na(res))) {
-    if (verbose) warning("Can't extract residuals from model.")
+    if (verbose) warning("Can't extract residuals from model.", call. = FALSE)
     res <- NULL
   } else if (yield_warning) {
-    warning(format_message(paste0("Can't extract '", res_type, "' residuals. Returning response residuals.")), call. = FALSE)
+    warning(format_message(paste0(
+      "Can't extract '", res_type, "' residuals. Returning response residuals."
+    )), call. = FALSE)
   }
 
   res
@@ -160,7 +162,7 @@ get_residuals.coxph <- function(x, weighted = FALSE, verbose = TRUE, ...) {
 #' @export
 get_residuals.crr <- function(x, weighted = FALSE, verbose = TRUE, ...) {
   if (isTRUE(weighted) && isTRUE(verbose)) {
-    warning("Weighted residuals are not supported for 'crr' models.", call. = FALSE)
+    warning("Weighted residuals are not supported for `crr` models.", call. = FALSE)
   }
   x$res
 }
@@ -185,7 +187,7 @@ get_residuals.slm <- function(x, weighted = FALSE, verbose = TRUE, ...) {
   )
 
   if (is.null(res) || all(is.na(res))) {
-    if (verbose) warning("Can't extract residuals from model.")
+    if (verbose) warning("Can't extract residuals from model.", call. = FALSE)
     res <- NULL
   }