+++ R/get_statistic.R
@@ -61,7 +61,7 @@ get_statistic.default <- function(x, column_index = 3, verbose = TRUE, ...) {
   # edge cases: check for NULL
   params <- rownames(cs)
   if (is.null(params)) {
-    params <- paste(1:nrow(cs))
+    params <- paste(seq_len(nrow(cs)))
   }
 
   out <- data.frame(
@@ -488,10 +488,7 @@ get_statistic.gamlss <- function(x, ...) {
 
 #' @export
 get_statistic.vglm <- function(x, ...) {
-  if (!requireNamespace("VGAM", quietly = TRUE)) {
-    stop("Package 'VGAM' needed for this function to work. Please install it.")
-  }
-
+  check_if_installed("VGAM")
   cs <- VGAM::coef(VGAM::summary(x))
 
   out <- data.frame(
@@ -563,7 +560,21 @@ get_statistic.cgam <- function(x,
 
 #' @export
 get_statistic.coxph <- function(x, ...) {
-  get_statistic.default(x, column_index = 4)
+  # z is not always in the same column
+  # not sure t is possible, but it is cheap to include it in the regex
+  # avoid calling default method which would be computationally wasteful, since
+  # we need summary() here.
+  cs <- suppressWarnings(stats::coef(summary(x)))
+  column_index <- grep("^z$|^t$|Chisq", colnames(cs))
+  out <- data.frame(
+    Parameter = row.names(cs),
+    Statistic = cs[, column_index, drop = TRUE],
+    stringsAsFactors = FALSE,
+    row.names = NULL
+  )
+  out <- text_remove_backticks(out)
+  attr(out, "statistic") <- find_statistic(x)
+  out
 }
 
 
@@ -854,7 +865,7 @@ get_statistic.multinom <- function(x, ...) {
 
   if (is.matrix(stderr)) {
     se <- c()
-    for (i in 1:nrow(stderr)) {
+    for (i in seq_len(nrow(stderr))) {
       se <- c(se, as.vector(stderr[i, ]))
     }
   } else {
@@ -1232,7 +1243,7 @@ get_statistic.btergm <- function(x, verbose = TRUE, ...) {
   bootstraps <- x@boot$t
 
   # standard error
-  sdev <- sapply(1:ncol(bootstraps), function(i) {
+  sdev <- sapply(seq_len(ncol(bootstraps)), function(i) {
     cur <- (bootstraps[, i] - params[i])^2
     sqrt(sum(cur) / length(cur))
   })
@@ -2360,4 +2371,14 @@ get_statistic.bfsl <- function(x, ...) {
   out <- text_remove_backticks(out)
   attr(out, "statistic") <- find_statistic(x)
   out
+}
+
+
+#' @export
+get_statistic.lm_robust <- function(x, ...) {
+  if (is_multivariate(x)) {
+    get_statistic.mlm(x, ...)
+  } else {
+    get_statistic.default(x, ...)
+  }
 }